Holy shit this be a huge huge commit

+++ app/links.php
@@ -44,16 +44,21 @@ $links = get_all_links();
                         </tr>
                         <tr>
                             <td style="padding: 10px;">
-                                <?php foreach ($links as $link): ?>
+                                <?php foreach ($links as $index =>$link): ?>
                                     <div class="link">
-                                        <p class="name"><?= $link["name"] ?></p>
+                                        <p class="name">
+                                            <?= $link["name"] ?>
+                                            <?php if ($index === 0): ?>
+                                                <img src="assets/symbols/boom.gif" alt="">
+                                            <?php endif; ?>
+                                        </p>
                                         <a href="clicklink.php?linkid=<?= $link["id"] ?>"><?= $link["target"] ?></a>
                                         <?php if ($link["description"]): ?>
                                             <p class="description"><?= $link["description"] ?></p>
                                         <?php endif; ?>
                                         <p class="stats">
-                                            <span class="clicks"><?= $link["visits"] ?> clicks</span> - Added by <?= $link["username"] ?>
-                                            <?php if($link["username"] == $session["username"]): ?>
+                                            <span class="clicks"><?= $link["visits"] ?> clicks</span> - Added by <?= $link['username'] === $session["username"] ? "you!" : $link["username"] ?>
+                                            <?php if($link["username"] == $session["username"] && !$session["admin"]): ?>
                                                 - <a href="deletelink.php?linkid=<?=$link['id']?>">Delete</a>
                                             <?php endif; ?>
                                         </p>
